+++ tencentcloud/service_tencentcloud_scf.go
@@ -89,14 +89,14 @@ func (me *ScfService) CreateFunction(ctx context.Context, info scfFunctionInfo) 
 		ratelimit.Check(request.GetAction())
 
 		if _, err := client.CreateFunction(request); err != nil {
-			return retryError(errors.WithStack(err), "InternalError")
+			return retryError(errors.WithStack(err))
 		}
 		return nil
 	}); err != nil {
 		return err
 	}
 
-	return waitScfFunctionReady(ctx, info.name, *info.namespace, client)
+	return nil
 }
 
 func (me *ScfService) DescribeFunction(ctx context.Context, name, namespace string) (resp *scf.GetFunctionResponse, err error) {